backend/requirements.txt
# Flask et extensions
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
Flask-Compress==1.14

# Base de données
cx-Oracle==8.3.0
SQLAlchemy==2.0.23
alembic==1.12.1

# Authentification et sécurité
python-ldap==3.4.3
PyJWT==2.8.0
bcrypt==4.1.2
cryptography==41.0.7

# Validation et sérialisation
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Intelligence artificielle
openai-whisper==20231117
piper-tts==1.2.0
ollama==0.1.7
torch==2.1.1
torchaudio==2.1.1

# Téléphonie
pyst2==0.5.1
requests==2.31.0

# Cache et performance
redis==5.0.1
celery==5.3.4

# Monitoring et observabilité
prometheus-client==0.19.0
structlog==23.2.0
python-json-logger==2.0.7

# Utilitaires
python-dotenv==1.0.0
pytz==2023.3
click==8.1.7
Werkzeug==2.3.7

# Tests
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Linting et qualité de code
black==23.11.0
flake8==6.1.0
mypy==1.7.1
bandit==1.7.5

# Documentation
flasgger==0.9.7.1 